requirements.txt
flask
pyodbc
python-dotenv
yfinance
pandas
numpy
pytz
openpyxl
json
requests
